requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
python-pptx==1.0.2
python-docx==1.1.2
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.13
groq==0.5.0
google-generativeai==0.7.2
